+++ app/me/bets/[year]/page.tsx
@@ -40,6 +40,9 @@ export default async function Page(props: Props) {
 
   const bets = await getBetsForPlayerCached(player.id, year.year);
 
+  // Go to presentation mode without group slug - let the presentation mode handle group selection
+  const presentationUrl = `/${year.year}/${Object.values(categories)[0].slug}`;
+
   return (
     <MainContainer>
       <div className={styles.header}>
@@ -50,14 +53,11 @@ export default async function Page(props: Props) {
         </Typography>
         <div className={styles.buttonWrapper}>
           <div className={styles.backLink}>
-            <Button element="a" href="/me">
-              &lt; Dashboard
+            <Button element="a" href="/me/bets">
+              &lt; My Bets
             </Button>
           </div>
-          <Button
-            element="a"
-            href={`/${year.year}/${Object.values(categories)[0].slug}`}
-          >
+          <Button element="a" href={presentationUrl}>
             Go to presentation mode &gt;
           </Button>
         </div>