+++ app/me/layout.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { redirect } from 'next/navigation';
-import YearList from 'components/dashboard/YearList';
 import { auth0 } from 'lib/auth0';
 import styles from './me.module.scss';
 
@@ -12,10 +11,5 @@ export default async function MeLayout({
   const session = await auth0.getSession();
   if (!session) redirect('/');
 
-  return (
-    <div className={styles.mainWrapper}>
-      <YearList />
-      {children}
-    </div>
-  );
+  return <div className={styles.mainWrapper}>{children}</div>;
 }